Fb oasis t+0 (#13)
In the original code a coupling lag had to be used, coupling took place at the last time step but not at the first, and the coupling fields had to be written in an oasis restart file. With this ticket, the program can run in a different way by not using a coupling lag, coupling taking place at the first time step but not the last, and the coupling fields are written in the wave restart file - no extra oasis restart file needed.

+++ regtests/ww3_tp2.14/input/prep_env.sh
@@ -78,7 +78,11 @@ ln -sf ../input/namcouple.$swtstr namcouple
 
 echo '   copy toy model inputs'
 cd $path_w
-cp ../input/toy/r-toy.nc.$swtstr r-toy.nc
+if [ -f ../input/toy/r-toy.nc.$swtstr ]; then
+  cp ../input/toy/r-toy.nc.$swtstr r-toy.nc
+else
+  echo "WARNING: model input ../input/toy/r-toy.nc.$swtstr does not exist"
+fi
 ln -sf ../input/toy/grid_toy_model.nc .
 ln -sf ../input/toy/toy_coupled_field.nc.$swtstr toy_coupled_field.nc
 ln -sf ../input/TOYNAMELIST.nam.$swtstr TOYNAMELIST.nam
@@ -86,7 +90,11 @@ ln -sf ../input/toy/toy_model .
 
 echo '   copy ww3 model inputs'
 cd $path_w
-cp ../input/r-ww3.nc.$swtstr r-ww3.nc
+if [ -f ../input/r-ww3.nc.$swtstr ]; then
+  cp ../input/r-ww3.nc.$swtstr r-ww3.nc
+else
+  echo "WARNING: model input ../input/toy/r-ww3.nc.$swtstr does not exist"
+fi
 cp ../input/ww3_shel_${swtstr}.inp ../input/ww3_shel.inp
 cp ../input/ww3_shel_${swtstr}.nml ../input/ww3_shel.nml
 echo ''
